requirements.txt
# Core dependencies for DinoV3 fine-tuning on TNG50-SKIRT

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# HuggingFace
transformers>=4.56.0
datasets>=2.14.0
accelerate>=0.20.0
peft>=0.7.0  # For LoRA

# Astronomy data handling
astropy>=5.0.0
astroquery>=0.4.6

# Image processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0

# Data utilities
tqdm>=4.65.0
pandas>=2.0.0

# Experiment tracking (optional)
wandb>=0.15.0
tensorboard>=2.13.0

# API and download
requests>=2.28.0
urllib3>=1.26.0

# Configuration
PyYAML>=6.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Development tools (optional)
jupyter>=1.0.0
ipython>=8.0.0
